Make the beacon dismissable on pages without an ApolloProvider
